pyproject.toml
[tool.poetry]
name = "integrations-challenge"
version = "0.1.0"
description = ""
authors = ["Felipe Armoni <felipe.armoni@gmail.com>"]
readme = "README.md"
packages = [{include = "integrations_challenge"}]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
requests = "^2.28.1"
python-dotenv = "^0.21.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
pylint = "^2.15.2"
black = "^22.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
